+++ HISTORY.md
@@ -24,6 +24,7 @@
 * objects: added category labels to unified palette
 * gui: accelerated unified palette scrolling animation
 * adjust scroll bars when refreshing the palette
+* store: commented out saving the unified palette setting in the project xml during development
 
 ### 2021-07-07
 * morphic, gui: tweaked perish() animation